refactor : 레거시 클래스 제거
2022. 02. 13
레거시 클래스 제거, 새로운 클래스로 대체함.

Resolver : loop-study
See also : heesunjeong
