--- src/test/java/nextstep/subway/unit/TokenAuthenticationInterceptorTest2.java
@@ -1,97 +1,0 @@
-package nextstep.subway.unit;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import nextstep.auth.authentication.AuthenticationConverter;
-import nextstep.auth.authentication.AuthenticationToken;
-import nextstep.auth.authentication.TokenAuthenticationInterceptor2;
-import nextstep.auth.authentication.UserDetailService;
-import nextstep.auth.context.Authentication;
-import nextstep.auth.token.JwtTokenProvider;
-import nextstep.auth.token.ObjectMapperBean;
-import nextstep.auth.token.TokenRequest;
-import nextstep.auth.token.TokenResponse;
-import nextstep.member.domain.LoginMember;
-import org.apache.http.HttpStatus;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.http.MediaType;
-import org.springframework.mock.web.MockHttpServletRequest;
-import org.springframework.mock.web.MockHttpServletResponse;
-
-import java.io.IOException;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.when;
-
-@ExtendWith(MockitoExtension.class)
-class TokenAuthenticationInterceptorTest2 {
-    private static final String EMAIL = "email@email.com";
-    private static final String PASSWORD = "password";
-    public static final String JWT_TOKEN = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIiLCJuYW1lIjoiSm9obiBEb2UiLCJpYXQiOjE1MTYyMzkwMjJ9.ih1aovtQShabQ7l0cINw4k1fagApg3qLWiB8Kt59Lno";
-
-    @Mock
-    private UserDetailService userDetailsService;
-
-    @Mock
-    private JwtTokenProvider jwtTokenProvider;
-
-    @Mock
-    private ObjectMapperBean objectMapper;
-
-    @Mock
-    private AuthenticationConverter converter;
-
-    @InjectMocks
-    private TokenAuthenticationInterceptor2 interceptor;
-
-    @Test
-    void authenticate() {
-        // given
-        when(userDetailsService.loadUserByUsername(EMAIL)).thenReturn(new LoginMember(1L, EMAIL, PASSWORD, 20));
-
-        // when
-        AuthenticationToken authenticationToken = new AuthenticationToken(EMAIL, PASSWORD);
-        Authentication authentication = interceptor.authenticate(authenticationToken);
-
-        // then
-        assertThat(authentication.getPrincipal()).isNotNull();
-    }
-
-    @Test
-    void preHandle() throws IOException {
-        // given
-        LoginMember member = new LoginMember(1L, EMAIL, PASSWORD, 20);
-
-        when(converter.convert(any())).thenReturn(new AuthenticationToken(EMAIL, PASSWORD));
-        when(userDetailsService.loadUserByUsername(EMAIL)).thenReturn(member);
-        when(objectMapper.writeValueAsString(any())).thenReturn(new ObjectMapper().writeValueAsString(new Authentication(member).getPrincipal()));
-        when(jwtTokenProvider.createToken(anyString())).thenReturn(JWT_TOKEN);
-        when(objectMapper.writeValueAsString(any())).thenReturn(new ObjectMapper().writeValueAsString(new TokenResponse(JWT_TOKEN)));
-
-        // when
-        MockHttpServletRequest request = createMockRequest();
-        MockHttpServletResponse response = new MockHttpServletResponse();
-        interceptor.preHandle(request, response, new Object());
-
-        // then
-        assertAll(
-                () -> assertThat(response.getStatus()).isEqualTo(HttpStatus.SC_OK),
-                () -> assertThat(response.getContentType()).isEqualTo(MediaType.APPLICATION_JSON_VALUE),
-                () -> assertThat(response.getContentAsString()).isEqualTo(new ObjectMapper().writeValueAsString(new TokenResponse(JWT_TOKEN)))
-        );
-    }
-
-    private MockHttpServletRequest createMockRequest() throws IOException {
-        MockHttpServletRequest request = new MockHttpServletRequest();
-        TokenRequest tokenRequest = new TokenRequest(EMAIL, PASSWORD);
-        request.setContent(new ObjectMapper().writeValueAsString(tokenRequest).getBytes());
-        return request;
-    }
-
-}
